python-cleanroom/packages/botas/pyproject.toml
[project]
name = "botas"
version = "1.0.0"
requires-python = ">=3.10"

dependencies = [
    "httpx>=0.27.0",
    "PyJWT>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
rules = ["E", "F", "W", "I"]